backend/requirement.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
python-multipart
passlib[bcrypt]
python-jose[cryptography]
aiofiles
cryptography
python-dotenv
python-dateutil

